Moved all styling files to styles folder. Removed components folder from LandingPage component *Updated all imports related to moved files.
