+++ src/components/dashboard/ApprovedTeacherList.js
@@ -18,7 +18,7 @@ import Avatar from "@material-ui/core/Avatar";
 // Internal Components
 import SideBar from "../shared/components/Sidebar/SideBar.js";
 
-import "./Dashboard.css";
+import "../styles/Dashboard.css";
 
 const useStyles = makeStyles((theme) => ({
   paper: {
