+++ src/components/mentors/MentorList.js
@@ -4,7 +4,7 @@ import MentorPublicPage from "./MentorPublicPage";
 import { firestore } from "../../config/fbConfig";
 import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
-import "./mentors.css";
+import "../styles/mentors.css";
 
 const useStyles = makeStyles(theme => ({
   paper: {
