+++ src/components/styles/SideBar.scss
