+++ src/components/styles/auth_SignUp.scss
