+++ src/components/styles/calendar_main.scss
