+++ src/components/styles/landingpage_styles.scss
