+++ src/components/styles/sidebar.css
